Moved logic to a new class SignatureVerifier.
