+++ identity-hub-did/src/main/java/org/eclipse/dataspaceconnector/identityhub/did/IdentityHubDidExtension.java
@@ -58,7 +58,8 @@ public class IdentityHubDidExtension implements ServiceExtension {
         }
 
         var client = new IdentityHubClientImpl(httpClient, new ObjectMapper());
-        var credentialsVerifier = new IdentityHubCredentialsVerifier(client, monitor, didPublicKeyResolver);
+        var signatureVerifier = new SignatureVerifier(didPublicKeyResolver, monitor);
+        var credentialsVerifier = new IdentityHubCredentialsVerifier(client, monitor, signatureVerifier);
         context.registerService(IdentityHubCredentialsVerifier.class, credentialsVerifier);
 
         monitor.info("Initialized Identity Hub DID extension");
